requirements.txt
python-telegram-bot
pymongo
dnspython
telethon
pymorphy2
